api/requirements.txt
flask-restplus<=0.12.1
psycopg2<=2.7.5